删除了Component继承的设计,reload不好处理,既然都是一类了没必要再继承处理,使用if else走不同的逻辑就好了

+++ CSharp/Game/Model/Component/FactoryComponent.cs
@@ -42,7 +42,7 @@ namespace Model
 
         public T Create(int configId)
         {
-            int type = (int) World.Instance.GetComponent<ConfigComponent>().Get<UnitConfig>(configId).Type;
+            int type = World.Instance.GetComponent<ConfigComponent>().Get<UnitConfig>(configId).Type;
             return this.allConfig[type].Create(configId);
         }
     }
